Exposed the 'from_one_raw_str' function for use in defining new header structs.
