+++ src/header/common/content_type.rs
@@ -1,6 +1,6 @@
 use header::Header;
 use std::fmt::{mod, Show};
-use super::from_one_raw_str;
+use super::util::from_one_raw_str;
 use mime::Mime;
 
 /// The `Content-Type` header.
